+++ recipes/common.rb
@@ -21,15 +21,39 @@
 # Install common components
 include_recipe node['mcollective']['recipes']['install_common']
 
-# directory for unpackaged plugins (extra mcollective libdir)
-remote_directory node['mcollective']['site_plugins'] do
-  source 'plugins'
-  owner 'root'
-  group 'root'
-  files_owner 'root'
-  files_group 'root'
-  mode "0755"
-  recursive true
+# Create a group permitted to read mcollective config
+group node['mcollective']['group'] do
+  members node['mcollective']['users']
+  action :create
+end
+
+# directories for unpackaged plugins (extra mcollective libdir)
+plugin_types = %w{agent audit connector data facts registration security simplerpc_authorization}
+plugin_types.each do |dir|
+  directory "#{node['mcollective']['site_plugins']}/#{dir}" do
+    owner 'root'
+    group 'root'
+    mode "0755"
+    recursive true
+  end
+end
+
+# install chef agent
+cookbook_file "#{node['mcollective']['site_plugins']}/agent/chef.rb" do
+  source "plugins/agent/chef.rb"
+  mode '0644'
+  if resource(:service => 'mcollective')
+    notifies :restart, "service[mcollective]"
+  end rescue Chef::Exceptions::ResourceNotFound
+  only_if { node['mcollective']['install_chef_agent?'] }
+end
+cookbook_file "#{node['mcollective']['site_plugins']}/agent/chef.ddl" do
+  source "plugins/agent/chef.ddl"
+  mode '0644'
+  if resource(:service => 'mcollective')
+    notifies :restart, "service[mcollective]"
+  end rescue Chef::Exceptions::ResourceNotFound
+  only_if { node['mcollective']['install_chef_agent?'] }
 end
 
 # directory for per-plugin configuration (plugin.d)
@@ -44,18 +68,31 @@ end
 # stomp connector
 template "#{node['mcollective']['plugin_conf']}/stomp.cfg" do
   source "plugin-stomp.cfg.erb"
+<<<<<<< HEAD
   mode 0755
+=======
+  owner 'root'
+  group node['mcollective']['group']
+  mode '0640'
+>>>>>>> upstream/master
   variables :stomp => node['mcollective']['stomp']
 end
 
 # activemq connector
 template "#{node['mcollective']['plugin_conf']}/activemq.cfg" do
   source "plugin-activemq.cfg.erb"
+<<<<<<< HEAD
   mode 0755
+=======
+  owner 'root'
+  group node['mcollective']['group']
+  mode '0640'
+>>>>>>> upstream/master
   variables :stomp => node['mcollective']['stomp']
 end
 
 # rabbitmq connector
+<<<<<<< HEAD
 cookbook_file "#{node['mcollective']['rabbit_connector']}/rabbitmq.rb" do
   source "plugins/connector/rabbitmq.rb"
   mode 0755
@@ -66,4 +103,32 @@ end
 # install gem open4
 gem_package "open4" do
   action :install
+=======
+template "#{node['mcollective']['plugin_conf']}/rabbitmq.cfg" do
+  source "plugin-rabbitmq.cfg.erb"
+  owner 'root'
+  group node['mcollective']['group']
+  mode '0640'
+  variables :stomp => node['mcollective']['stomp'],
+            :rabbitmq => node['mcollective']['rabbitmq']
+end
+
+# redis connector
+if node['mcollective']['connector'] == 'redis'
+  # install the connector
+  gem_package "redis"
+  remote_file "#{node['mcollective']['site_plugins']}/connector/redis.rb" do
+    source "https://github.com/ripienaar/mc-plugins/raw/master/connector/redis/redis.rb"
+    mode 0644
+  end
+
+  # install plugin configuration
+  template "#{node['mcollective']['plugin_conf']}/redis.cfg" do
+    source "plugin-redis.cfg.erb"
+    user 'root'
+    group node['mcollective']['group']
+    mode '0640'
+    variables :redis => node['mcollective']['redis']
+  end
+>>>>>>> upstream/master
 end